backend/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==23.1
django-cors-headers==4.2.0
djoser==2.2.0
django-cors-headers==4.2.0
flake8==6.0.0
gunicorn==21.1.0
python-dotenv==1.0.0
Pillow==10.0.0
drf-extra-fields==3.6.1
reportlab==4.0.4
psycopg2-binary==2.9.3